# server/docs/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  dto.DetailErrorDto:
    properties:
      field:
        type: string
      message:
        type: string
      value:
        type: string
    type: object
  dto.ErrorDto:
    properties:
      codeMessage:
        type: string
      details:
        items:
          $ref: '#/definitions/dto.DetailErrorDto'
        type: array
      message:
        type: string
    type: object
  dto.PaginationDto:
    properties:
      page:
        type: integer
      pageCount:
        type: integer
      perPage:
        type: integer
      total:
        type: integer
    type: object
  dto.SummaryResponseDto:
    properties:
      balance:
        type: number
      expense:
        type: number
      income:
        type: number
    type: object
  dto.TransactionDto:
    properties:
      amount:
        type: number
      category:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      type:
        type: string
      updatedAt:
        type: string
    type: object
  dto.TransactionRequestDto:
    properties:
      amount:
        type: number
      category:
        type: string
      description:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
  dto.TransactionResponseDto:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.TransactionDto'
        type: array
      pagination:
        $ref: '#/definitions/dto.PaginationDto'
    type: object
host: localhost:8080
info:
  contact:
    email: kauebertaze2004@gmai.com
    name: KauêTTS
  description: API do sistema de controle financeiro
  title: Controle Financeiro API
  version: "1.0"
paths:
  /health:
    get:
      description: Retorna o status de funcionamento da API
      responses:
        "200":
          description: OK
      summary: Verifica se a API está online
      tags:
      - Health
  /v1/summary:
    get:
      description: Cálculo automático de entradas, saídas e saldo total
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SummaryResponseDto'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorDto'
      summary: Retorna o resumo financeiro
      tags:
      - Summary
  /v1/transactions:
    get:
      description: Visualização completa das movimentações financeiras cadastradas
      parameters:
      - description: 'Buscar por título ou descrição '
        in: query
        name: search
        type: string
      - description: Tipo da transação
        enum:
        - income
        - expense
        in: query
        name: type
        type: string
      - description: Categoria da transação
        in: query
        name: category
        type: string
      - default: 1
        description: Página atual
        in: query
        name: page
        type: integer
      - default: 10
        description: Quantidade de registros por página
        in: query
        name: perPage
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TransactionResponseDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorDto'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorDto'
      summary: Lista todas as transações
      tags:
      - Transaction
    post:
      description: Cadastro de receitas e despesas
      parameters:
      - description: Dados da transação
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TransactionRequestDto'
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorDto'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorDto'
      summary: Cadastrar transações
      tags:
      - Transaction
  /v1/transactions/{id}:
    delete:
      description: Remoção de transações cadastradas
      parameters:
      - description: ID da transação
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorDto'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorDto'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorDto'
      summary: Deletar transação
      tags:
      - Transaction
    put:
      description: Atualização de informações das transações
      parameters:
      - description: ID da transação
        in: path
        name: id
        required: true
        type: integer
      - description: Dados atualizados da transação
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TransactionRequestDto'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorDto'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorDto'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorDto'
      summary: Editar transação
      tags:
      - Transaction
produces:
- application/json
schemes:
- http
swagger: "2.0"